requirements.txt
APScheduler==3.7.0
bidict
certifi
charset-normalizer
dnspython
idna==3.4
mongoengine
numpy
pymongo
python-engineio==4.4.0
python-socketio==5.1.0
pytz==2022.7.1
rdx==0.0.4
requests
shapely
six==1.16.0
tzlocal==2.1
urllib3==1.26.15
websocket-client
fastapi[all]
uvicorn
fastapi-utils
lxml
xlsxwriter==3.1.1
dateutils==0.6.12
openpyxl
schedule
